introduce: set on boot implementation
This introduces the set on boot feature. This may
also require new permissions (boot_complete). All
settings inside Aero Control are now saved as long
as they are preferences.

One thing that might not work out of the box are
governor specific parameters. However, this can
be considered as very good start for a clean set-on-
boot option.

There are also some random fixes here and there.