sane warning dialogs in logcat
When the app is started, there is also a refresher Thread
started to update the values. While it stops refreshing
when another fragment is visible it doesn't check if the
app is running in forground at all. This could lead to
battery drain and also spam in logcat.

The next step would be to freeze those threads in the
meantime completely and wake them up, when the app
is visible again.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>