boot: show alert if random reboot was detected
Use the new boot receiver to show an alert if
the kernel detected a sudden reboot. It informs
the user to checkout /proc/last_kmsg
This requires a 3.0.x kernel, since our old kernel
didn't support this feature.

This could be even enhanced in the future with an
automatic copy to sdcard or something similar.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>