cpu: change governor specific menu into button
Before this patch the user had to change the governor to
see the governor specific parameters. If one just want to
change one tunable of the already set governor one has to
always set the governor again.
This changes this behaviour. I introduced a new action bar
button which gives the user the same functionality with one
less click.
Also fixing some bugs regarding permissions and increase
execution time by simple not sleeping that long.

setGovernor() is a generic method to set a governor for
all available cpus.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>