cpu: fix delay for governor settings
On the defy the delay would be just to short to
update everything correctly. The good news; we don't
need this delay on the defy. Since we can set permissions
correctly here, we can get governors parameter directly
without any hack.

Note: This was only tested on a 3.0 Kernel build.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>