statistic: redesign listview element
We can now safely use our own array adapter to accomplish a much better
effect. With this change some parts of the underlaying data structures
are sligthly changed which has the benefit that we can customize each
textview element independent.
Loading an array with multiple nested arrays has never been easier.

Keep in mind that the listview data and the holograph data are now
different (in theory).