memory: fix potential reboot while trimming
If one would trim partitions such as /system or /data
and cancle the actual dialog the trim-process is still
alive and keeps running in background (not the Wrapper-Thread).

This could cause problems if the user starts to trim right
before another trim process hasn't finished yet.

By making the dialog not cancelable the user has to wait
until it has finished, which is the right approach, after
a small sleep the dialog exits automatically.

A much better approach would be to actually use async-tasks for
such work, however this might be a little bit overkill.

Thanks to @Quarx2k who send me this bug :)

Signed-off-by: Blechd0se <alex.christ@hotmail.de>