misc: add vibration control
This adds vibration control to a new slider fragment called misc
settings. While we have the defy parts on our defy, we can actually give
other devices a special role too. Device specific features should be
placed here.

Notice: This also adds new permissions to Aero Control (VIBRATE). The
small PreferenceHanlder takes now an argument to force a virbration
after a change. While its surely useful for the vibration control, we
could also use it to inform the user that the input was not correct or
something.