shell: add a small workqueue api
When the users applies a profile in the profiles panel
or when the user has activated the set on boot functionality
setRootInfo() would be called a few times. While this makes
no problems on powerful devices, it can create small UI lags
on low end devices.

This patch introduces some sort of global workqueue where
all commands can be stored and executed in one row, which
means the su-process is called just ones and not serval times,
thus improving peformance.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>