per-app: add the backbone of per-app profiles (1/3)
This adds the backbone componentes for the per-app profiles.
perAppHelper creates for each available profile a new per-app
profile. This contains the selected applications as well as
system states (should system apps are visible or not).

PerAppService is the actual service behind all this. During app
changes it scans the foreground app and checks if there is
any profile out there which contains a match. If so, it will
set the new settings with help of the settingsHelper (future commit).
It also handles the default case handling (returning to previous state).

PerAppServiceHelper manages the service itself and starts it during
app start up. It can also cleanup everything if the user decides to
disable per-app profiles.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>