aero: add per-app handling into main process (3/3)
This adds the service handling into the main aero control
process. During startup it checks if the user wants to use
per-app profiles and starts the according service.
If changed later in the app settings it cleans up everything
set by itself.
This commit also contains the main "default" profile after
the system returns to a normal state.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>