memory: fix trimming on non-supported fs
If the filesystem in question doesn't support the trim
command it would just "loop" in the dialog until the
user forces to close the app.

Fix that by actually checking if the filesystem *maybe*
supports trim and show an options menu depending on the
fs.

The only thing this doesn't fix is checking for "fstrim"
if its supported on the filesystem, because even we are
on the right fs, it could just not be supported.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>