profiles: partially fix incredible long load
If many apps are installed it takes quite a time to gather all
information we need. Too long for my own taste; on a Nexus 4
its a average time of 6.5 seconds for nearly every profile.
Even the sorting will be faster with the second and following
iteration, it can easily lead up to 10 seconds.

Fix this partially by copying the list over to the next
profiles, so they can already use a sorted list.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>