shell: fix min/max cpu freq for sure
Keep it simple, stupid. Trying to get the frequency for
offlined cores is pointless at this point of time. The
previous implementation just didn't work. Instead of messing
around with the hotplug driver or other things, we can just
set the freq for all cores to the one we have on cpu0.

Its simple, it works and it won't create overhead because of
expansive su-calls.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>