profiles: greatly improve the per-app handling
Instead of calling the expansive methods directly on
the fragment creation we should only call it when the
user actually clicks on the small icon.
Because of one previous commit we can afterwards copy
the data to all other profiles and speed up the loading
process quite a bit.

The only bottleneck is now the actual sorting process,
anyway with this change the time taken on a Nexus 4 goes
down to aroud ~1 seconds.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>