aero: fix ISE during sub-fragment switch
If the user enters a sub-fragment and pulls out the navigation
drawer chooses a parent fragment and then hits the back button
Aero Control would just crash.
This is due an IllegalStateException inside the fragment switch
code, because the fragment in question was already added some
time ago.
Fix this by just letting the back counter logic decide what to
do and most importantly; don't crash.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>