helper: choose input type dynamically
When using the generic preference handler implementation
we always set the input type to numbers only. In most cases
this is fine, since we are mostly tweaking numbers. However
it would be nice to have the full keyboard if we set something
like arrays or other stuff (example: interactive gonveror tunables).

Signed-off-by: Blechd0se <alex.christ@hotmail.de>