cpu: add generic live oc/uc
Instead of using hardcoded values this interface
features now a fully generic backend. No matter how
many frequencies are found, it will adjust accordingly.

Since this was done from scratch there are also a ton of
code optimizations inside as well as less buggy code.

Since profiles/set-on-boot is independent from this we
don't have to touch it directly.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>