boot: implement a smarter set-on-boot feature (1/3)
Previously if a certain value has changed it would have got written
to our shared preference and the user would have no clue he has
changed a value. Only if one adds a profile the change would be
visible.
Now giving the user full control over the desired value by giving a
checkbox for each visible element indicating whether the value is
set on boot or not.

In this commit we enhanced the PreferenceHandler to allow this new
feature, which is now fully generic.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>