profiles: show spinner while loading app-data
Sorting the app data takes some time and until its finished
there shouldn't be any UI-interactions. Hence @Quarx2k suggested
to run the expensive operation in a seperated thread and update
the elements later in the UI-thread.

This allows us also to show the dialog again, after it was closed
when the user switches the system-apps.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>