splashscreen: add a new SplashScreen for new users
When starting Aero Control for the very first time this
splashscreen will occur guiding the user through the very
basics of this app.
We are actually adding value behind this, since we can set
up some basic things.

The SplashScreen acitivity now handles the start of Aero Control
in each case, which prevents a nasty su-PopUp questions right
after Aero Control was started the first time.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>