profiles: resolve null pointer if service is down
If the service helper, not the service itself, has no reference
we would end up in a null pointer. Since its actually okay if
the helper class is not instantiated we don't need to try-catch
it.
Also adding a warning boolean which fixes strange overlay issues
if the same fragment is created several times.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>