helpers: add the filemanager backbone (1/3)
This commit series introduces a new feature to aero
control called 'Your Settings'. This will allow the
user to set any desired option for every tunable
on the system.

Starting with the backbone, a new generic simple file
manager is added. It uses the root-calls only as some
sort of fallback and displays each folder/file found
on the system.
It's designed to be generic to be added to any view,
so it can be reused in the feature.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>