misc: add 'your settings' layout/logic (2/3)
As follow up to the previous commit this adds the
user interface for the new feature as well as the
basic logic operations inside the miscFragment.

Desired tunables can now be set via the new file
manager and modified with the normal PreferenceHandler.

The only missing thing is the set on boot feature,
which actually only requires a loop through the new
storage file and set the desired data since the other
things are already handled.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>