service: prepare perapp for android L
At and above android L getRunningTasks() will be
deprecated due to androids new document focused UI.
Since it will only return some sort of pseudo-data
per app profiles become useless.

But we are smarter than google and check for all apps
which are flagged with the IMPORTANCE_FOREGROUND flag.
Since the actual visible app is the first one in that
list, take it.

Since L is not officially released, we might want to
check this later again.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>