profiles: refactor perapp selection dialog
This should be a much cleaner solution for the user
than the previous one. Basically we are switching
from the android default multi-select handler to a
custom one which should be more flexible.
The good part is that the logic itself hasn't been
touched, so everything should work as before except
there's a new UI.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>