helpers: add interactive shell extension
Shell commands are very expansive. Normally they are queued
up as much as possible and executed afterwards. Sometimes its
easier to have an interactive shell, where we can simply pass
our commands to.
This commit adds a very simple interative shell, which will be
kept alive as long as desired. For now there are two places
where we can benefit from this:
  1.) Display color - where we can preview an instant result
      as soon as the user touches the seekbar without performance
      loss.
  2.) settingsHelper - This handles both the set-on-boot as well
      as the "normal" profile functionality. We are using the inbuilt
      Workqueue to fill our interactive shell.

In theory, we can expand the usage to a few more places. A big thanks
to @Chainfire who wrote the concept of this shell.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>