appmonitor: add the framework of appmonitor (1/3)
This adds the main framework of a new feature called appmonitor.
It allows to monitor app usage statistics on a per app basis.

It is designed to be very efficient and flexible, every monitor
point is a separate module (called AppModule). Those modules can
be loaded/unloaded during runtime. Each passed app gets an AppContext
which stores the main app usage data.
It has support for on/off state as well as a sleep modus to save power.

The whole process is designed to be power efficient in general. On a
Motorola Defy (with three loaded modules) the process takes about 10ms
to finish.
Because this framework is made to be called every couple of seconds,
it has also a small cleanup feature.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>