service/adapter: add service relevant changes for appmonitor (2/3)
This changes the per app service to make use of the new appmonitor
framework.
It also changes how the service is called in general; since API 19
google decided to change the AlarmManager behaviour (probably in
preparation for Android Ms Doze feature). Since its not anymore
garantueed to get a call every 5 seconds, we switch to a much more
simple solution.

This commit also introduces a fast way to read system files for the
shellHelper class.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>