appmonitor: add import/export of current data
One issue with saving all the data inside of java structures is the
potential data loss if the app dies.
This commit addresses this issue. If onDestroy() is called in our
main activity, we start an export of the current data in a json format.
As soon as the service restarts itself we import the saved data and
put it back to the right place (AppContext, Modules, MetaData, usw...).

Because onDestroy() shouldn't be called when we power off the device,
we don't need to care about logic when we reboot the phone.
This commit also introduces a few more APIs to save data directly and
without the framework.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>