helpers: add helptexts to various tunables
This commit introduces helptexts for various tunables. The actual goal
is to support as much tunables as possible to help users understand the
respective tunable.

Initially, this commit supports upto 70 different tunable helptexts,
which can be extended in the future and for different languages quite
easily via the provided hashmap.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>