perapp: use UsageStatsManager for lollipop+
Since lollipop, we can't just get the current active/visible app like before.
Google introduced UsageStatsManager to allow access to the usage stats.

A nasty sideeffect is, that the user actually needs to enable it in the settings
for each app individually. So this commit extends the framework slightly to
support this new method.

We still need to make this same process work when the user starts the app for
the first time (tutorial fragments) and enabled the PerAppService there.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>