appmonitor: redesign detail fragment of appmonitor
This commit aims for a redesign of the detail fragment from appmonitor.
It actually looks a lot more like the material concept from google.

In the same breath i enhanced a few core functionalities to make further
changes simpler and easier to maintain.
This probably also fixes a few bugs here and there.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>