appmonitor: fix import/export of data and create only one thread
Earlier, we created one thread on onCreate() and onStartCommand() when
we started the service. Now its only one (as intended).

Secondly, we export the data now on a regular basis (default 60 seconds)
and import it back in if the service was killed in the meantime.
This should make the framework a little bit more robust.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>