shellHelper: refactor shell-handling
Befor this commit, each su-call was really expansive. It created
a new process[shell] for each respective call. Since this creates
a unnecessary overhead, we create only one shell instance per app
lifecycle. The only exception is the boot/profile service, which
get a second one to handle race conditions better.

Overall, this should make the app performance quite a bit snappier
on key points. The new shell-framework is flexible enough to be
extended in the future.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>