helpers: material: checkbox: partially fix high cpu load
This checkbox implementation seems to cause a massive cpu load
when viewed. The actual problem is, that invalidate() is called
a gazillion times per second to draw it, even its already drawn.

This commit partially fixes this problem, by removing the high load
when the checkbox is not checked. It still causes a moderate high
load if the checkbox is already checked.

Thanks to @Quarx2k for debugging and the heads up.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>