appmonitor: save the right data on exportData()
We actually saved the output from module.getValues() which could
contain any data, but not the one we wanted. Since we can access
the raw data via getRawData() for the respective module, the fix
is quite simple.
This should fix strange behaviour after Aero Control was closed
and the data for all apps looked the same.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>